test(loading): remove redundant tests (#27430)
Issue number: N/A

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

There were several redundant tests I missed during the initial generator
pass.

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->


https://github.com/ionic-team/ionic-framework/pull/27430/commits/b8a3df049e42eb1fb206f2391f35bddfe7200b4e

- Moved the `htmlAttributes` e2e test to a spec test


https://github.com/ionic-team/ionic-framework/pull/27430/commits/fb3f96840c04d3c7310e9c896bdca77964caaa99

- Removed the HTML content screenshot test. HTML content is already
tested in `test/loading.spec.ts`.


https://github.com/ionic-team/ionic-framework/pull/27430/commits/fc7c166e0e715f6976b0c137debd4b3d82752569

- Removed the per-direction checks on the non-basic tests as these
behaviors do not vary across directions. The basic test still has
per-direction checks
- Also removed the per-mode check on the translucent test since this
feature is available on iOS only


https://github.com/ionic-team/ionic-framework/pull/27430/commits/8524b6c9d607d786a500a908fab6b770ad118cc6

- The isOpen tests are testing that you can open/close the modal by
triggering the `isOpen` property. I combined the two tests into 1 test
since they are doing the same thing: Checking that the modal opens when
`isOpen` is `true` and closes when `isOpen` is false. This also avoids
the 500ms timeout of the 2nd test which should speed up execution.


https://github.com/ionic-team/ionic-framework/pull/27430/commits/f6dac8b47c93765085bc32cbbba9a5f78d156ea6

- Removed the "standalone" screenshot in favor of just checking that the
overlay is visible. Rendering of the contents of the overlay are not
dependent on `ion-app` being present.


https://github.com/ionic-team/ionic-framework/pull/27430/commits/a2d628910170a28882e2589831c40db8dfb5f2f2

- The `runVisualTests` util in `test/basic` was checking that every
loading instance could also dismiss. This is only needed for the basic
test, so I removed this logic from the shared util and add it just to
the basic test.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->